+++ x/ibchooker/types/key.go
@@ -1,0 +1,5 @@
+package types
+
+const (
+	ModuleName = "ibchooker"
+)
