--- x/ibctransferhooks/types/key.go
@@ -1,5 +1,0 @@
-package types
-
-const (
-	ModuleName = "ibctranferhooks"
-)
